requirements.txt
httpx
httpx-curl-cffi
lxml
zendriver
latest-user-agents
user-agents
python-dotenv
